Move getting telemetry into its own file
